refactor(modal): 다시 (modal) 디렉터리 사용
