+++ app/layout.tsx
@@ -4,7 +4,7 @@ import Board from "@/components/Board";
 import Aside from "@/components/Aside";
 import Footer from "@/components/Footer";
 import Popular from "@/components/Popular";
-import Modal from "@/components/Modal";
+import Modal from "@/components/(modal)/Modal";
 import { FC, PropsWithChildren, ReactNode } from "react";
 import ScrollButton from "@/components/ScrollButton";
 import { generateOpenGraph } from "@/utils";
