+++ components/(modal)/Modal/index.tsx
@@ -2,7 +2,7 @@
 
 import { useAtomValue } from "jotai";
 import { modalContext } from "@/context/index";
-import { background } from "./style.css";
+import { background } from "../style.css";
 
 const Modal = () => {
   const modal = useAtomValue(modalContext);
